third_party/blink/web_tests/http/tests/devtools/sources/debugger-ui/function-generator-details-expected.txt

Tests that Debugger.getGeneratorObjectDetails command returns correct result.


Running: testiterNotStarted
iterNotStarted: type = object, subtype = generator
lineNumber = 22
columnNumber = 19
script is valid: yes
[[GeneratorState]] = suspended
[[GeneratorFunction]] = function* gen()
      {
          yield 1;
          yield 2;
          yield 3;
      }
[[GeneratorReceiver]] = Window

Running: testiterSuspended
iterSuspended: type = object, subtype = generator
lineNumber = 24
columnNumber = 10
script is valid: yes
[[GeneratorState]] = suspended
[[GeneratorFunction]] = function* gen()
      {
          yield 1;
          yield 2;
          yield 3;
      }
[[GeneratorReceiver]] = Window

Running: testiterClosed
iterClosed: type = object, subtype = generator
lineNumber = 22
columnNumber = 19
script is valid: yes
[[GeneratorState]] = closed
[[GeneratorFunction]] = function* gen()
      {
          yield 1;
          yield 2;
          yield 3;
      }
[[GeneratorReceiver]] = Window

Running: testiterObjGenerator
iterObjGenerator: type = object, subtype = generator
lineNumber = 33
columnNumber = 14
script is valid: yes
[[GeneratorState]] = suspended
[[GeneratorFunction]] = function*()
          {
              yield 11;
              yield 12;
              yield 13;
          }
[[GeneratorReceiver]] = Object

Running: testanonymousGenIter
anonymousGenIter: type = object, subtype = generator
lineNumber = 47
columnNumber = 10
script is valid: yes
[[GeneratorState]] = suspended
[[GeneratorFunction]] = function*() {
          yield 21;
          yield 22;
          yield 23;
      }
[[GeneratorReceiver]] = Window
